+++ clients/device-amqp/src/main/java/org/eclipse/hono/client/device/amqp/EventSender.java
@@ -13,25 +13,38 @@
 
 package org.eclipse.hono.client.device.amqp;
 
+import io.opentracing.SpanContext;
 import io.vertx.core.Future;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.proton.ProtonDelivery;
 
 /**
  * A client for sending event messages to Hono's AMQP adapter.
+ *
+ * @see <a href="https://www.eclipse.org/hono/docs/user-guide/amqp-adapter/#publishing-events">AMQP Adapter User Guide</a>
  */
-public interface EventSender extends AmqpSenderLink {
+public interface EventSender {
 
     /**
-     * Sends a event message for a given device.
+     * Sends an event message.
      *
-     * @param deviceId The id of the device.
      * @param payload The data to send.
      *            <p>
-     *            The payload will be contained in the message as an AMQP 1.0 <em>Data</em> section.
-     * @param contentType The content type of the payload (may be {@code null}).
+     *            The payload will be added to the message as an AMQP 1.0 <em>Data</em> section.
+     * @param contentType The content type of the payload or {@code null} if unknown.
      *            <p>
      *            This parameter will be used as the value for the message's <em>content-type</em> property.
+     * @param tenantId The tenant that the device belongs to or {@code null} to determine the tenant from 
+     *                 the device that has authenticated to the AMQP adapter.
+     *                 Unauthenticated clients must provide a non-{@code null} value to indicate the tenant of the
+     *                 device that the message originates from.
+     * @param deviceId The identifier of the device that the message originates from or {@code null} if the
+     *                 message originates from the device that has authenticated to the AMQP adapter.
+     *                 Authenticated gateway devices can use this parameter to send a message on behalf of
+     *                 another device.
+     *                 Unauthenticated clients must provide a non-{@code null} value to indicate the device that
+     *                 the message  originates from.
+     * @param context The OpenTracing context to use for tracking the execution of the operation (may be {@code null}).
      * @return A future indicating the outcome of the operation.
      *         <p>
      *         The future will succeed if the message has been accepted (and settled) by the peer.
@@ -41,10 +54,13 @@ public interface EventSender extends AmqpSenderLink {
      *         the future will be failed with either a {@code org.eclipse.hono.client.ServerErrorException} or a
      *         {@link org.eclipse.hono.client.ClientErrorException} depending on the reason for the failure to
      *         process the message.
-     * @throws NullPointerException if any of device-id or payload is {@code null}.
+     * @throws NullPointerException if payload is {@code null}.
+     * @throws IllegalArgumentException if tenant ID is not {@code null} but device ID is {@code null}.
      */
-    Future<ProtonDelivery> send(
-            String deviceId,
+    Future<ProtonDelivery> sendEvent(
             Buffer payload,
-            String contentType);
+            String contentType,
+            String tenantId,
+            String deviceId,
+            SpanContext context);
 }